+++ include/fractol_sets.h
@@ -6,7 +6,7 @@
 /*   By: ladloff <ladloff@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/28 21:54:55 by ladloff           #+#    #+#             */
-/*   Updated: 2023/05/28 23:07:04 by ladloff          ###   ########.fr       */
+/*   Updated: 2023/07/26 13:56:27 by ladloff          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,12 +16,12 @@
 # include "fractol.h"
 
 /* mandelbrot.c */
-void	mandelbrot(t_complex c, int *iter, int max_iter);
+void	mandelbrot(t_complex c, double *iter, int max_iter);
 
 /* julia.c */
-void	julia(t_complex *z, t_complex c, int *iter, int max_iter);
+void	julia(t_complex *z, t_complex c, double *iter, int max_iter);
 
 /* burning_ship.c */
-void	burning_ship(t_complex c, int *iter, int max_iter);
+void	burning_ship(t_complex c, double *iter, int max_iter);
 
 #endif /* FRACTOL_SETS_H */
